// src/sdks/Newcomputation/computationV2ApiDocs.ts
export type getComputationDataSourceListProps = {
  computationDataId: number;
  pageNum: number;
  pageSize: number;
};